Refactor order to be in service.  Refactor broadcast receiver to be in service.

+++ Apphangar/src/main/java/ca/mimic/apphangar/Tasks.java
@@ -16,10 +16,11 @@ public class Tasks extends SQLiteOpenHelper {
     public static final String COLUMN_LAUNCHES = "launches";
     public static final String COLUMN_TIMESTAMP = "timestamp";
     public static final String COLUMN_BLACKLISTED = "blacklisted";
-    public static final String COLUMN_ORDER = "order";
+    public static final String COLUMN_ORDER = "sort_order";
+    public static final String COLUMN_WIDGET_ORDER = "sort_widget_order";
 
     private static final String DATABASE_NAME = "tasks.db";
-    private static final int DATABASE_VERSION = 5;
+    private static final int DATABASE_VERSION = 7;
 
     // Database creation sql statement
     private static final String DATABASE_CREATE = "create table "
@@ -31,7 +32,9 @@ public class Tasks extends SQLiteOpenHelper {
             + " integer not null default 0, " + COLUMN_LAUNCHES
             + " integer not null default 1, " + COLUMN_TIMESTAMP
             + " datetime not null default current_timestamp, " + COLUMN_BLACKLISTED
-            + " bool not null default 0);";
+            + " bool not null default 0, " + COLUMN_ORDER
+            + " integer not null default 0, " + COLUMN_WIDGET_ORDER
+            + " integer not null default 0);";
 
     public Tasks(Context context) {
         super(context, DATABASE_NAME, null, DATABASE_VERSION);
@@ -47,6 +50,11 @@ public class Tasks extends SQLiteOpenHelper {
         Log.w(Tasks.class.getName(),
                 "Upgrading database from version " + oldVersion + " to "
                         + newVersion + ", which will destroy all old data");
+        if (newVersion >= 6 && oldVersion <= 5) {
+            db.execSQL("ALTER TABLE " + TABLE_TASKS + " ADD COLUMN " + COLUMN_ORDER + " INTEGER DEFAULT 0");
+            db.execSQL("ALTER TABLE " + TABLE_TASKS + " ADD COLUMN " + COLUMN_WIDGET_ORDER + " INTEGER DEFAULT 0");
+            return;
+        }
         db.execSQL("DROP TABLE IF EXISTS " + TABLE_TASKS);
         onCreate(db);
     }
